% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_fct.R
\name{load_data}
\alias{load_data}
\title{load datafile}
\usage{
load_data(filepath)
}
\arguments{
\item{filepath}{filepath}
}
\value{
dataframe
}
\description{
Load a dataframe from an excel sheet and call EndoPaste on it
}
